Rename 'Preconditions' to 'Assert'
Per team meeting, last week.

GearVRf-DCO-1.0-Signed-off-by: Jon Shemitz
j.shemitz@samsung.com
